Fix encoding issues in the loader
The issues I have experienced were related to `encodeURIComponent`
function not encoding `!` and `'`. Moving to caching the style bodies in
a map resolves that since we no longer need to encode / decode.

This also removes the wall of text generated from the webpack hack used
to load the style.

I added a comment in define to trigger the error during `yarn build`.

Closes: #240 (Build issues with font-properties in SFC)
Closes: #239 (Module not found: Error: Can't resolve 'important[...] in SFC)