add support for unmounting tags
This way the user can cleanly remove tags that have been added for
example in the case of modals or context menus.

This came out from discussions with @sombee on how to currently handle
those cases. This won't necessarily be the canonical way to do it but
since we already expose mount, having an equivalent way of disposing
tags makes sense.