Move to mono-repo (#631)
The various parts of imba are very intertwined. For convenience we are moving the typescript-imba-plugin and vscode-imba repositories into the main imba repository. This will (in combination with lerna) make it much easier to publish new version of the tooling and make sure the interdependent packages stay in sync.

One "downside" is that we've had to move the main imba files into packages/imba. This does, however, make it easier to split the compiler and the runtime at a later time. Will also probably move the imba.io site into the main repository because that too depends on both imba and the typescript-imba-plugin to generate certain docs.

We should reorganize the package.json scripts and improve the readme to make it much easier to get started with contributing to imba and the related packages.