Improve router
Using sessionStorage to use introspectable history stack so that you can look back/forward in the history at all times. When route is changed - the request now includes `apply` and `revert` arrays containing all the states between the current state and the previous one. Still need to update typings for router and events, and document the additions.